Added support for printing persistence indices in dimension 0
The idea is to be able to link back the information supplied in the
persistence pairing to an actual data set. For this it is *crucial*
to know which simplex gives rise to which feature.

I am not entirely satisfied with the way this is solved; subsequent
iterations of this should probably only return the *creators*.